fix(project): 修复MIDI轨道索引数据类型溢出问题
- 将track_index字段从u8改为u16以支持更多轨道
- 更新project_io模块中的轨道索引转换逻辑
- 确保原始MIDI轨道索引映射功能正常工作

+++ crates/yinhe-egui/src/project_io.rs
@@ -344,7 +344,7 @@ pub fn midi_to_archive_with_names(
                 uuid,
                 name,
                 color: [0.5, 0.5, 0.5],
-                track_index: new_idx as u8,
+                track_index: new_idx as u16,
                 channel_prefix: new_track_channel_prefixes[new_idx],
             });
         }
@@ -375,7 +375,7 @@ pub fn midi_to_archive_with_names(
             uuid,
             name,
             color: [0.5, 0.5, 0.5],
-            track_index: original_idx as u8,
+            track_index: original_idx as u16,
             channel_prefix: midi.track_channel_prefixes.get(original_idx).copied().flatten(),
         });
     }
